✨ fix #50: add soft-delete support 🔮

+++ api/src/actors.py
@@ -5,12 +5,13 @@ from fastapi.responses import JSONResponse
 from sqlmodel import (
     Field,
     SQLModel,
+    UniqueConstraint,
     Session,
     select,
     JSON,
     Relationship,
-    UniqueConstraint,
     desc,
+    null,
 )
 from sqlalchemy import Column, DateTime, func
 from typing import Iterable, Dict, Optional
@@ -37,12 +38,15 @@ class Actor(SQLModel, table=True):
     updateTimestamp: Optional[datetime] = Field(
         sa_column=Column(DateTime(timezone=True), onupdate=func.now(), nullable=True),
     )
+    deletedTimestamp: Optional[datetime] = Field(
+        sa_column=Column(DateTime(timezone=True), nullable=True),
+    )
     labels: Dict[str, str] = Field(default={}, sa_type=JSON)
     annotations: Dict[str, str] = Field(default={}, sa_type=JSON)
 
 
 class ActorsResult(SQLModel):
-    count: int | None = None # fix types
+    count: int | None = None  # fix types
     items: Iterable[Actor] | None = None
 
 
@@ -78,6 +82,9 @@ def read_actors(
         countSelect = countSelect.where(Actor.namespace_name == namespace_name)
         itemsSelect = itemsSelect.where(Actor.namespace_name == namespace_name)
 
+    countSelect = countSelect.where(Actor.deletedTimestamp == null())
+    itemsSelect = itemsSelect.where(Actor.deletedTimestamp == null())
+
     result = ActorsResult()
     result.count = session.exec(countSelect).one()
     result.items = session.exec(itemsSelect).all()
@@ -94,6 +101,7 @@ def read_latest_actors(
     statement = select(Actor)
     if namespace_name is not None:
         statement = statement.where(Actor.namespace_name == namespace_name)
+    statement = statement.where(Actor.deletedTimestamp == null())
     statement = statement.order_by(desc(Actor.creationTimestamp))
     statement = statement.offset(offset).limit(limit)
     return session.exec(statement).all()
@@ -106,6 +114,7 @@ def read_actor(
     statement = select(Actor)
     statement = statement.where(Actor.namespace_name == namespace_name)
     statement = statement.where(Actor.name == actor_name)
+    statement = statement.where(Actor.deletedTimestamp == null())
     return session.exec(statement).one()
 
 
@@ -132,10 +141,19 @@ def update_actor(
 def delete_actor(
     namespace_name: str, actor_name: str, session: Session = Depends(get_session)
 ) -> JSONResponse:
-    # or how can we run a delete query directly?
+    namespace = read_namespace(namespace_name)
     actor = read_actor(namespace_name, actor_name)
-    session.delete(actor)
-    session.commit()
+
+    softDelete = namespace.labels.get("soft-delete") == "true"
+
+    if softDelete:
+        actor.deletedTimestamp = datetime.now()
+        actor.annotations = {}
+        session.add(actor)
+        session.commit()
+    else:
+        session.delete(actor)
+        session.commit()
     return JSONResponse(
         status_code=status.HTTP_202_ACCEPTED, content={"message": "Actor deleted"}
     )
